requirements.txt
nox
distro
urllib3
requests
lxml
python-dateutil
pyOpenSSL
